refactor(settings): share font family helpers
